% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistGDM.R
\name{DistGDM}
\alias{DistGDM}
\title{Compute the environmental or biological distance between pairs of sites}
\usage{
DistGDM(site.grid, gdm.rast, env.vars, dist = "mhtn")
}
\arguments{
\item{site.grid}{A two column data.frame with the cell id of pairs of sites to compare. Cell ids must match the env.vars raster* object.}

\item{gdm.rast}{A gdm model object.}

\item{env.vars}{A rasterStack object of gdm transformed environmental predictors. Obtained by using the gdm.transform option in the gdm package.}

\item{dist}{Either "bio" to compute the biological distance between pairs of sites or "mhtn" to compute the manhattan environmental distance between pairs of sites.}
}
\value{
A vector of biological distances between pairs of sites.
}
\description{
Compute the environmental or biological distance between pairs of sites
}
