% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadGBIF.R
\name{DownloadGBIF}
\alias{DownloadGBIF}
\title{Download GBIF data}
\usage{
DownloadGBIF(key, user, user.email, pwd, country, custom.shp,
  n.max = 2e+06, subset.size = 2e+06)
}
\arguments{
\item{key}{a gbif taxon key obtained through the rgbif::name_backbone function. Required}

\item{user}{your GBIF account user name. Required.}

\item{user.email}{your GBIF account email. Required.}

\item{pwd}{your GBIF password. Required.}

\item{country}{ISO2 code of the coutry to download data from.}

\item{custom.shp}{a SpatialPolygons* object to download data from. Ignored if country is not NULL.}

\item{n.max}{integer. The maximum number of records before triggering the subset option.}

\item{subset.size}{integer. The number of records to subset randomly from file}
}
\value{
a list with:
\describe{
  \item{occ.table}{Occurrence table for the defined taxon key and area}
  \item{citation}{The citation of your GBIF download}
}
}
\description{
This function retrieves GBIF data in SIMPLE_CSV for a user provided taxon key and area, the latter either as a ISO2
country code or a SpatialPolygons* object in WGS84 (EPSG: 4326) projection.
}
\note{
By default this function will sample at random 2 million records for files
with more than 2 milion rows. Change the default n.max and subset.size to change
this behavior.
}
\examples{

\dontrun{
library(raster)
library(rgbif)
library(WhereNext)

#Example 1: Get occurrence data for colombian bats by country code
gbif.key <- name_backbone(name = "Chiroptera")
gbif.res <- DownloadGBIF(gbif.key$orderKey, "your username", "your email", "your password", "CO") #Enter your GBIF credentials here

#Example2: Get occurrence data for bats from the Antioquia province in Colombia by using a shapefile
col.shp <- getData(name="GADM",country="CO", level=0) #Get country shapefile with level 1 administrative boundaries
antioquia <- col.shp[2, ]
gbif.res <- DownloadGBIF(gbif.key$orderKey, "your username", "your email", "your password", NULL, antioquia)
}
}
