% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapGDMLight.R
\name{MapGDMLight}
\alias{MapGDMLight}
\title{Create a PCA transformed set of variables to map GDM}
\usage{
MapGDMLight(rast.trans)
}
\arguments{
\item{rast.trans}{A RasterStack of gdm transformed environmental predictors returned by the gdm.transform function in gdm-package}
}
\value{
A list with the following objects:
\describe{
  \item{pcaRast}{A RasterStack with the first three principal components}
  \item{pcaRast.all}{A RasterStack with the all the principal components}
}
}
\description{
An internal function to create a PCA transformed set of variables to map spatially the results of GDM analysis using an RGB palette.
}
\examples{
gdm.map <- MapGDMLight(rast.trans)
plotRGB(gdm.map)
}
