% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemCorrLayers.R
\name{RemCorrLayers}
\alias{RemCorrLayers}
\title{Remove correlated layers}
\usage{
RemCorrLayers(in.stack, t.cor)
}
\arguments{
\item{in.stack}{a RasterStack object}

\item{t.cor}{the maximum allowed correlation between layers in the RasterStack. Must be a value between 0 and 1.}
}
\value{
A RasterStack of layers with correlation below the user specified correlation threshold.
}
\description{
For a pair of correlated raster layers, this function will remove one. The output stack will have layers with a correlation between them below the user specified threshold.
}
\examples{
env.vars.uncorr <- RemCorrLayers(env.vars, 0.8)
}
