% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunWhereNext.R
\name{RunWhereNext}
\alias{RunWhereNext}
\title{Run WhereNext}
\usage{
RunWhereNext()
}
\description{
This function starts the shiny application WhereNext
}
\note{
This implementation of WhereNext is limited to occurrence files of up to 2 million rows.
If greater, it will sample at random up to 2 million rows in unix based systems or none in Windows.
Only up to 100.000 records selected at random will be shown in the map tab.
}
\examples{
WhereNext()

}
\author{
Jorge Velásquez-Tibatá
}
