% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleCSV.R
\name{SampleCSV}
\alias{SampleCSV}
\title{Sample rows from a text file}
\usage{
SampleCSV(file, percORn, nrLines, sep = "\\t", dec = ".", mxPerc = 1)
}
\arguments{
\item{file}{The full path to the file to be read}

\item{percORn}{the proportion (if less than 1) or the number of rows to be sampled from the file.}

\item{nrLines}{integer. If known, the number rows in the file.}

\item{sep}{the field separator character. See ?read.delim for details.}

\item{dec}{the character used in the file for decimal points. See ?read.delim for details.}

\item{mxPerc}{maximum proportion (0 to 1) of rows to extract.}
}
\value{
a data frame with a subset of rows of the original file
}
\description{
This function samples either a percentage or a number of rows from a text file at random without loading the entire file into R, thus reducing memory usage. Useful to get a sample of very large files. Modified from package DMwR2 by Luis Torgo ltorgo@dcc.fc.up.pt.
}
\examples{
occs <- SampleCSV("largeFile.csv", 10000)
}
