% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhereNext-package.R
\docType{package}
\name{WhereNext-package}
\alias{WhereNext-package}
\alias{WhereNext}
\title{WhereNext: A recommending system for biological surveys}
\description{
WhereNext is a recommendation system to identify iteratively sites that complement the most existing biodiversity surveys. WhereNext will guide you through the process of developing a General Dissimilarity Model (GDM) for your group and area of interest based on occurrence data as well as identifying iteratively survey sites that maximize your potential to record new species.
}
\details{
\tabular{ll}{
Package: \tab WhereNext\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2019-07-16\cr
License: \tab GNU 3.0\cr
}
}
\references{
Vel치squez-Tibat치, J. (2019) \emph{WhereNext}: A recommending system for biological surveys. Version 1.0.0
}
\author{
\strong{Jorge Vel치squez-Tibat치}\cr
email: \email{jivelasquezt@gmail.com}
}
\keyword{package}
